Reorder and add color to the case summary text
