+++ src/components/modecules/ConfirmedCasesSummary.js
@@ -2,11 +2,13 @@ import React from "react"
 import { useStaticQuery, graphql } from "gatsby"
 import { useTranslation } from "react-i18next"
 import { Label } from "../atoms/Text"
+import { Box } from "@material-ui/core"
 import styled from "styled-components"
+import { mapColorForStatus } from "../../utils/colorHelper"
 
-const BoldLabel = styled(Label)`
+const SummaryBox = styled(Box)`
    {
-    font-weight: 700;
+    margin: 10px 0px;
   }
 `
 
@@ -28,18 +30,29 @@ const ConfirmedCasesSummary = props => {
 
   const { t } = useTranslation()
 
+  const statusOrdering = {
+    deceased: 10,
+    critical: 20,
+    serious: 30,
+    hospitalised: 40,
+    discharged: 50,
+  }
+
   return (
-    <React.Fragment>
-      <BoldLabel display="inline">
-        {t("confirmed_case_summary.status_summary")}：
-      </BoldLabel>
-      <Label display="inline">
-        {status
-          .sort((a, b) => b.totalCount - a.totalCount)
-          .map(v => t(`cases.status_${v.fieldValue}`) + "：" + v.totalCount)
-          .join("　")}
-      </Label>
-    </React.Fragment>
+    <SummaryBox>
+      {status
+        .sort(
+          (a, b) => statusOrdering[a.fieldValue] - statusOrdering[b.fieldValue]
+        )
+        .map(v => (
+          <Label
+            display="inline"
+            style={{ color: mapColorForStatus(v.fieldValue).main }}
+          >
+            {t(`cases.status_${v.fieldValue}`)}：{v.totalCount}　
+          </Label>
+        ))}
+    </SummaryBox>
   )
 }
