+++ src/components/organisms/CaseCard.js
@@ -8,90 +8,13 @@ import { withLanguage } from "../../utils/i18n"
 import { Label } from "@components/atoms/Text"
 import { DefaultChip } from "@components/atoms/Chip"
 import {
-  pink,
-  teal,
-  blueGrey,
-  red,
-  green,
-  amber,
-  grey,
-} from "@material-ui/core/colors"
+  mapColorForClassification,
+  mapColorForStatus,
+} from "../../utils/colorHelper"
 import { formatDateDDMM } from "@/utils"
 import MuiLink from "@material-ui/core/Link"
 import * as d3 from "d3"
 const colors = d3.scaleOrdinal(d3.schemeDark2).domain([0, 1, 2, 3, 4])
-
-const mapColorForClassification = classification => {
-  const mapping = {
-    imported: {
-      main: pink[600],
-      contrastText: "#fff",
-    },
-    imported_close_contact: {
-      main: pink[700],
-      contrastText: "#fff",
-    },
-    local: {
-      main: teal[600],
-      contrastText: "#fff",
-    },
-    local_possibly: {
-      main: teal[500],
-      contrastText: "#fff",
-    },
-    local_unknown_source: {
-      main: teal[700],
-      contrastText: "#fff",
-    },
-    local_possibly_close_contact: {
-      main: teal[400],
-      contrastText: "#fff",
-    },
-    local_close_contact: {
-      main: teal[500],
-      contrastText: "#fff",
-    },
-    default: {
-      main: grey[900],
-      contrastText: "#fff",
-    },
-  }
-
-  if (!mapping[classification]) return mapping["default"]
-  return mapping[classification]
-}
-
-const mapColorForStatus = status => {
-  const mapping = {
-    hospitalised: {
-      main: amber[700],
-      contrastText: "#000",
-    },
-    discharged: {
-      main: green[700],
-      contrastText: "#fff",
-    },
-    serious: {
-      main: red[600],
-      contrastText: "#fff",
-    },
-    critical: {
-      main: red[900],
-      contrastText: "#fff",
-    },
-    deceased: {
-      main: blueGrey[800],
-      contrastText: "#fff",
-    },
-    default: {
-      main: grey[50],
-      contrastText: "#000",
-    },
-  }
-
-  if (!mapping[status]) return mapping["default"]
-  return mapping[status]
-}
 
 const WarsCaseContainer = styled(Box)`
   background: ${props =>
